fix data-science editor
